Filtering: Remove DSPFilters.
Now filtering section implements TPT-transformed SVF.
Shoutout to Vadim Zavalishin (for original work introducing TPT) &
@JordanTHarris, who wrote the original JUCE code.

+++ generators/generator.cpp
@@ -89,7 +89,7 @@ void Generator::set_freq(float8 _freq)
     phase_inc = freq * freq_cst;
 }
 
-double Generator::get_phase()
+float8 Generator::get_phase()
 {
     return phase;
 }
